refactor: address review feedback - move files and use gems_assert_equal
- Move sparse_attn_indexer.py from fused/DSA/ to fused/ directory
- Move test from tests/test_DSA/ to tests/ directory
- Use gems_assert_equal for strict equality checks in tests
- Update import paths in __init__.py and benchmark
