+++ src/flag_gems/fused/__init__.py
@@ -4,7 +4,7 @@ from flag_gems.fused.concat_and_cache_mla import concat_and_cache_mla
 from flag_gems.fused.cross_entropy_loss import cross_entropy_loss
 from flag_gems.fused.cutlass_scaled_mm import cutlass_scaled_mm
 from flag_gems.fused.DSA.bin_topk import bucket_sort_topk
-from flag_gems.fused.DSA.sparse_attn_indexer import sparse_attn_indexer
+from flag_gems.fused.sparse_attn_indexer import sparse_attn_indexer
 from flag_gems.fused.FLA import (
     chunk_gated_delta_rule_fwd,
     fused_recurrent_gated_delta_rule_fwd,
